ADD support for torchscript (mostly type annotations, initializing objects), REMOVE timing code (time.time() incompatible with torchscript

+++ src/data/det_utils.py
@@ -203,7 +203,7 @@ class ConvertCoco(object):
         # Store info in target dict
         target['area'] = area
         target['person_id'] = person_id_arr.tolist()
-        target['image_size'] = [w, h]
+        target['image_size'] = torch.FloatTensor([w, h])
         target['id'] = ids.tolist()
         target['iou_thresh'] = iou_thresh
         target['is_known'] = is_known
